# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_joaofilippe_subclub_internal_web_common.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  internal_web_auth.LoginRequestDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_web_auth.TokenResponseDTO:
    properties:
      token:
        type: string
    type: object
  internal_web_customer.AddressDTO:
    properties:
      city:
        type: string
      complement:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      state:
        type: string
      street:
        type: string
      zipCode:
        type: string
    type: object
  internal_web_customer.CustomerDTO:
    properties:
      active:
        type: boolean
      address:
        $ref: '#/definitions/internal_web_customer.AddressDTO'
      createdAt:
        type: string
      document:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  internal_web_customer.CustomerInputDTO:
    properties:
      active:
        type: boolean
      address:
        $ref: '#/definitions/internal_web_customer.AddressDTO'
      document:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  internal_web_customer.PaginatedCustomerResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_web_customer.CustomerDTO'
        type: array
      totalCount:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend application for SubClub
  title: SubClub API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns a signed JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_auth.LoginRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_web_auth.TokenResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Authenticate
      tags:
      - auth
  /api/v1/customers:
    get:
      description: Get a paginated list of customers
      parameters:
      - description: Search by name or email
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_customer.PaginatedCustomerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: List customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Creates a new customer in the system
      parameters:
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/internal_web_customer.CustomerInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_customer.CustomerDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Create a customer
      tags:
      - customers
  /api/v1/customers/{id}:
    delete:
      description: Soft deletes a customer by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Delete a customer
      tags:
      - customers
    get:
      description: Retrieves a customer by their UUID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_customer.CustomerDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Get a customer by ID
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Updates an existing customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/internal_web_customer.CustomerInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_customer.CustomerDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Update a customer
      tags:
      - customers
...
swagger: "2.0"